change adu file size to be uint64_t in case of larger updates

+++ sdk/inc/azure/iot/az_iot_adu_client.h
@@ -383,7 +383,7 @@ typedef struct
   /**
    * Size of a file, in bytes.
    */
-  uint32_t size_in_bytes;
+  uint64_t size_in_bytes;
   /**
    * Hashes provided for a given file in the update request.
    */
